Split TermWriter into TermWriter and TextTermWriter
