+++ org.spoofax.terms/src/org/spoofax/terms/io/TermWriter.java
@@ -2,37 +2,38 @@ package org.spoofax.terms.io;
 
 import org.spoofax.interpreter.terms.IStrategoTerm;
 
+import java.io.ByteArrayOutputStream;
 import java.io.IOException;
 import java.io.OutputStream;
-import java.io.PrintWriter;
 
 /**
  * Interface for term writers.
+ *
+ * A term writer writes a term to some textual or binary format.
+ * Configuration parameters on the term writer determine how the terms are written,
+ * such as whether to use indentation in a textual writer,
+ * or whether to use maximal sharing in a binary writer.
+ *
+ * @implSpec Textual term writers should implement the {@link TextTermWriter} interface.
+ * @see TextTermWriter
+ * @see SimpleTextTermWriter
  */
 public interface TermWriter {
 
     /**
-     * Gets a singleton instance of the default {@link TextTermWriter} implementation.
-     *
-     * @return the singleton instance
-     */
-    static TextTermWriter getInstance() { return TextTermWriter.INSTANCE; }
-
-    /**
-     * Writes the term to a string.
+     * Writes the term to a byte array.
      *
      * @param term the term to write
-     * @return the string representation of the term
+     * @return the byte representation of the term
      */
-    default String writeToString(IStrategoTerm term) {
-        StringBuilder sb = new StringBuilder();
-        try {
-            write(term, sb);
+    default byte[] writeToBytes(IStrategoTerm term) {
+        try (ByteArrayOutputStream stream = new ByteArrayOutputStream()) {
+            write(term, stream);
+            return stream.toByteArray();
         } catch (IOException e) {
-            // This couldn't have happened!
-            throw new RuntimeException("Unexpected exception: ", e);
+            // The ByteArrayOutputStream implementation never throws an IOException.
+            throw new RuntimeException("Unexpected exception: " + e.getMessage(), e);
         }
-        return sb.toString();
     }
 
     /**
@@ -42,19 +43,6 @@ public interface TermWriter {
      * @param outputStream the output stream
      * @throws IOException an I/O exception occurred
      */
-    default void write(IStrategoTerm term, OutputStream outputStream) throws IOException {
-        try (PrintWriter writer = new PrintWriter(outputStream)) {
-            write(term, writer);
-        }
-    }
-
-    /**
-     * Writes the term to the specified appendable.
-     *
-     * @param term the term to write
-     * @param writer the appendable
-     * @throws IOException an I/O exception occurred
-     */
-    void write(IStrategoTerm term, Appendable writer) throws IOException;
+    void write(IStrategoTerm term, OutputStream outputStream) throws IOException;
 
 }
